+++ UnitTests/SocketSettingsTest.cs
@@ -8,7 +8,7 @@ namespace UnitTests
     public class SocketSettingsTest
     {
         [Test]
-        public void TestSocketSettingsDefaultValues()
+        public void DefaultValues()
         {
             SocketSettings socketSettings = new SocketSettings();
 
@@ -28,8 +28,7 @@ namespace UnitTests
             Assert.IsTrue(socketSettings.RequireClientCertificate);
         }
 
-        [Test]
-        public void TestSocketSettingsConfigure()
+        private Dictionary BaseTestDict()
         {
             Dictionary dict = new Dictionary();
             dict.SetBool(SessionSettings.SOCKET_NODELAY, false);
@@ -46,9 +45,14 @@ namespace UnitTests
             dict.SetBool(SessionSettings.SSL_CHECK_CERTIFICATE_REVOCATION, false);
             dict.SetBool(SessionSettings.SSL_ENABLE, true);
             dict.SetBool(SessionSettings.SSL_REQUIRE_CLIENT_CERTIFICATE, false);
+            return dict;
+        }
 
+        [Test]
+        public void Configure()
+        {
             SocketSettings socketSettings = new SocketSettings();
-            socketSettings.Configure(dict);
+            socketSettings.Configure(BaseTestDict());
 
             Assert.IsFalse(socketSettings.SocketNodelay);
             Assert.AreEqual(1, socketSettings.SocketReceiveBufferSize);
@@ -64,6 +68,38 @@ namespace UnitTests
             Assert.IsFalse(socketSettings.CheckCertificateRevocation);
             Assert.IsTrue(socketSettings.UseSSL);
             Assert.IsFalse(socketSettings.RequireClientCertificate);
+        }
+
+        private SocketSettings CreateWithSslConfig(bool sslValidateCertificates, bool sslCheckCertificateRevocation)
+        {
+            var dict = BaseTestDict();
+            dict.SetBool(SessionSettings.SSL_VALIDATE_CERTIFICATES, sslValidateCertificates);
+            dict.SetBool(SessionSettings.SSL_CHECK_CERTIFICATE_REVOCATION, sslCheckCertificateRevocation);
+            SocketSettings socketSettings = new SocketSettings();
+            socketSettings.Configure(dict);
+            return socketSettings;
+        }
+
+        [Test]
+        public void Configure_SslOverride()
+        {
+            // SSLValidateCertificates=true does not affect SSLCheckCertificateRevocation
+            var socketSettings = CreateWithSslConfig(true, false);
+            Assert.IsTrue(socketSettings.ValidateCertificates);
+            Assert.IsFalse(socketSettings.CheckCertificateRevocation);
+
+            socketSettings = CreateWithSslConfig(true, true);
+            Assert.IsTrue(socketSettings.ValidateCertificates);
+            Assert.IsTrue(socketSettings.CheckCertificateRevocation);
+
+            // SSLValidateCertificates=false means SSLCheckCertificateRevocation must be false
+            socketSettings = CreateWithSslConfig(false, true);
+            Assert.IsFalse(socketSettings.ValidateCertificates);
+            Assert.IsFalse(socketSettings.CheckCertificateRevocation);
+
+            socketSettings = CreateWithSslConfig(false, false);
+            Assert.IsFalse(socketSettings.ValidateCertificates);
+            Assert.IsFalse(socketSettings.CheckCertificateRevocation);
         }
     }
 }